requirements.txt
Django==4.0
django-extensions==3.1.5
Faker==10.0.0
ipython==7.30.1

django-crispy-forms~=1.13.0